test: Reduce test flakiness
