+++ pyproject.toml
@@ -16,7 +16,8 @@ test = [
 	"pytest-sugar~=1.0",
 	"pytest~=8.0",
 	"nox~=2024.10.9",
-	"pytest-xdist>=3.6.1",
+	"pulumi-random>=4.18.0",
+	"pytest-order>=1.3.0",
 ]
 policy = ["pulumi-policy>=1,<2"]
 aws = ["pulumi-aws>=6"]
@@ -60,10 +61,6 @@ explicit_package_bases = true
 
 [tool.pytest.ini_options]
 addopts = [
-	# ? Some tests mutually exclusive and require fresh environments.
-	# ? So for now, we need at minimum 2 processes (see `tests/test_resource_.py` file).
-	"--numprocesses=2",
-	"--dist=loadgroup",
 	"--strict-markers",
 	"--doctest-modules",
 	"--cov",
