+++ tests/_helpers.py
@@ -1,0 +1,5 @@
+class Order:
+    """Test run order keys for `pytest.mark.order`."""
+
+    BeforeRegistryInit = 100
+    AfterRegistryInit = 200
