+++ tests/test_resource_.py
@@ -4,25 +4,19 @@ import pytest
 
 from pulumi_extra import get_resource_cls, resource_has_attribute
 from pulumi_extra.errors import UnknownResourceTypeError
-
-
-@pytest.fixture(autouse=True)
-def reset_cache() -> None:
-    """Reset cache for each test."""
-    resource_has_attribute.cache_clear()
-    get_resource_cls.cache_clear()
+from tests._helpers import Order
 
 
 class Test__resource_has_attribute:
-    @pytest.mark.xdist_group("registry-initialized")
+    @pytest.mark.order(Order.AfterRegistryInit)
     def test(self) -> None:
         # Arrange
-        import pulumi_aws  # noqa: F401
+        import pulumi_random  # noqa: F401
 
         # Act & Assert
-        assert resource_has_attribute("aws:s3/bucket:Bucket", "bucket") is True
-        assert resource_has_attribute("aws:s3/bucket:Bucket", "acl") is True
-        assert resource_has_attribute("aws:s3/bucket:Bucket", "my-attribute") is False
+        assert resource_has_attribute("random:index/randomId:RandomId", "byte_length") is True
+        assert resource_has_attribute("random:index/randomId:RandomId", "keepers") is True
+        assert resource_has_attribute("random:index/randomId:RandomId", "my-attribute") is False
 
     def test_unknown_resource_type(self) -> None:
         # Arrange
@@ -31,36 +25,36 @@ class Test__resource_has_attribute:
         # Act & Assert
         with pytest.raises(
             UnknownResourceTypeError,
-            match="Unable to resolve resource type 'aws:unknown/unknown:Unknown'",
+            match="Unable to resolve resource type 'random:unknown/unknown:Unknown'",
         ):
-            resource_has_attribute("aws:unknown/unknown:Unknown", "whatever")
+            resource_has_attribute("random:unknown/unknown:Unknown", "whatever")
 
 
 class Test__get_resource_cls:
-    @pytest.mark.xdist_group("registry-initialized")
+    @pytest.mark.order(Order.AfterRegistryInit)
     def test(self) -> None:
         # Arrange
-        import pulumi_aws  # noqa: F401
+        import pulumi_random  # noqa: F401
 
         # Act
-        cls = get_resource_cls("aws:accessanalyzer/analyzer:Analyzer")
+        cls = get_resource_cls("random:index/randomId:RandomId")
 
         # Assert
         assert cls is not None
-        assert f"{cls.__module__}.{cls.__name__}" == "pulumi_aws.accessanalyzer.analyzer.Analyzer"
+        assert f"{cls.__module__}.{cls.__name__}" == "pulumi_random.random_id.RandomId"
 
-    @pytest.mark.xdist_group("fresh-env")
+    @pytest.mark.order(Order.BeforeRegistryInit)
     def test_registry_not_initialized(self) -> None:
         """If registry not initialized, it will return `None`."""
         # Arrange
         # ...
 
         # Act & Assert
-        assert get_resource_cls("aws:accessanalyzer/analyzer:Analyzer") is None
+        assert get_resource_cls("random:index/randomId:RandomId") is None
 
     def test_unknown_resource_type(self) -> None:
         # Arrange
         # ...
 
         # Act & Assert
-        assert get_resource_cls("aws:unknown/unknown:Unknown") is None
+        assert get_resource_cls("random:unknown/unknown:Unknown") is None
